+++ paramdb/_param_data/_dataclasses.py
@@ -12,11 +12,16 @@ from paramdb._param_data._param_data import ParamData
 
 @dataclass_transform(kw_only_default=True)
 class _ParamDataclass(ParamData):
-    """Base class for parameter dataclasses."""
+    """
+    Base class for parameter dataclasses.
+
+    Any keyword arguments given when creating a subclass are passed to the dataclass
+    constructor.
+    """
 
     def __init_subclass__(cls, /, kw_only: bool = True, **kwargs: Any) -> None:
         # Convert subclasses into dataclasses
-        super().__init_subclass__()
+        super().__init_subclass__()  # kwargs are passed to dataclass constructor
         dataclass(kw_only=kw_only, **kwargs)(cls)
 
     def __getitem__(self, name: str) -> Any: